update ActiveUsers component file structure
